+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Support special case of centering a nested shape with default x/y within a parent with default width/height calculated from nested shape
 - Consider Canvas Shape DSL support for LineAttributes `line_dash_offset` and `line_miter_limit`
 - Canvas Shape DSL Polygon `include?` does an outer/inner check of edge detection only
+- Ensure all Canvas Shape DSL properties are restored upon painting a shape to what they were prior to painting that shape
 
 ### 4.18.5.4